requirements.txt
Jinja2==2.7.3
requests==2.5.2
Frozen-Flask==0.11
pytz==2014.10

# These are dependencies of the above, but are specificied explicitly in order to pin versions


